+++ ocaml-lsp-server/test/e2e/__tests__/ocamllsp-hoverExtended.ts
@@ -56,7 +56,16 @@ describe("ocamllsp/hoverExtended", () => {
     expect(result).toMatchObject({
       contents: {
         kind: "plaintext",
-        value: outdent`int -> int -> int -> int -> int -> int -> int -> int -> int -> int`,
+        value: outdent`int ->
+         int ->
+         int ->
+         int ->
+         int ->
+         int ->
+         int ->
+         int ->
+         int ->
+         int`,
       },
       range: {
         end: { character: 5, line: 0 },